Conflicts:
	tests/travis/test

+++ setup.py
@@ -472,6 +472,9 @@ path = abspath(dirname(__file__))
 
 print()
 if 'develop' in sys.argv:
+    if exists('bokeh/__conda_version__.py'):
+        print(bright(red("ERROR:")) + " Detected a __conda_version__.py file, exiting")
+        sys.exit(1)
     check_remove_bokeh_install(site_packages)
     with open(path_file, "w+") as f:
         f.write(path)
@@ -581,6 +584,7 @@ setup(
         'bokeh.util.tests',
     ],
     package_data={'bokeh': package_data},
+    entry_points={'console_scripts': ['bokeh = bokeh.__main__:main',], },
     author='Continuum Analytics',
     author_email='info@continuum.io',
     url='http://github.com/bokeh/bokeh',
